bug: store score as int32_t instead of int16_t
Given the search depth I started notifying that the "good moves"
starting being extremely bad. This could ofc mean overflow which was in
fact the case.
I was a bit naive with my previous implementation. Instead increase the
scoring to use 32bit value to avoid over/underflows.
Also printing the score from the evaluator always returning 0. Also
fixed that bug.

+++ src/attack_generation.h
@@ -28,7 +28,7 @@ namespace {
         King   100    200    300    400    500    600
 */
 
-using MvvTableEntry = std::array<int16_t, 12>;
+using MvvTableEntry = std::array<int32_t, 12>;
 constexpr auto s_mvvLvaTable = std::to_array<MvvTableEntry>(
     {
         { 105, 205, 305, 405, 505, 605, 105, 205, 305, 405, 505, 605 },
@@ -46,7 +46,7 @@ constexpr auto s_mvvLvaTable = std::to_array<MvvTableEntry>(
     });
 }
 
-constexpr static inline int16_t getMvvLvaScore(Piece attacker, Piece victim)
+constexpr static inline int32_t getMvvLvaScore(Piece attacker, Piece victim)
 {
     return s_mvvLvaTable.at(attacker).at(victim);
 }
